+++ internals/Cargo.toml
@@ -15,12 +15,13 @@ exclude = ["tests", "contrib"]
 
 [features]
 default = []
-std = ["alloc"]
-alloc = []
+std = ["alloc", "hex/std"]
+alloc = ["hex/alloc"]
 
 test-serde = ["serde", "serde_json", "bincode"]
 
 [dependencies]
+hex = { package = "hex-conservative", version = "0.2.0", default-features = false }
 serde = { version = "1.0.103", default-features = false, optional = true }
 
 # Don't enable these directly, use `test-serde` feature instead.
