+++ primitives/README.md
@@ -3,6 +3,11 @@
 This crate provides primitive data types that are used throughout the
 [`rust-bitcoin`](https://github.com/rust-bitcoin) ecosystem.
 
+## Semver compliance
+
+Functions marked as unstable (e.g. `foo__unstable`) are not guaranteed to uphold semver compliance.
+They are primarily provided to support `rust-bitcoin`.
+
 ## Minimum Supported Rust Version (MSRV)
 
 This library should always compile with any combination of features on **Rust 1.63.0**.
